The alg.trees.Tree is now generic.
Several changes happened with this commit:

1. Made the `TreeLike` trait generic, it now requires two types. In
other words, `TreeLike[A, N]` where the type `A` represents the values
that are going to be stored in the `nodes` and `N` represents the type
associated with the calculation of the tree's amplitude.

2. The `TreeLike` trait doesn't expose anymore the method `amplitud`
since the method `maxAmplitud` doesn't need it anymore. From an API
perspective it didn't made sense since the only parameter as a `List[A]`
and not a `Tree[A]`.

3. The `TreeLike.findPaths` method doesn't return anymore paths with one
single leaf-node since they are not considered paths. Adding them
initially was a mistake.

4. We are setting the stage for the following methods.

```
  def size(tree: Tree[A]): Int = ???
  def max(t: Tree[A])(implicit ord: Ordering[A]): A = ???
  def depth: (Int, List[A]) = ???
  def map[B](f: (A) => B): Tree[B] = ???
```